SauvolaNet/requirements.txt
absl-py==0.13.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
argh==0.26.2
asn1crypto==1.3.0
astroid==2.4.2
astropy==4.0.1.post1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.9.1
bitarray==1.4.0
bkcharts==0.2
bleach==3.1.5
bokeh==2.1.1
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
clang==5.0
click==7.1.2
cloudpickle==1.5.0
clyent==1.2.2
colorama==0.4.3
contextlib2==0.6.0.post1
contextvars==2.4
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.21
cytoolz==0.10.1
dask==2.20.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distributed==2.20.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
flatbuffers==1.12
fsspec==0.7.4
future==0.18.2
gast==0.4.0
gevent==20.6.2
glob2==0.7
gmpy2==2.0.8
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.40.0
h5py==3.1.0
HeapDict==1.0.1
html5lib==1.1
idna==2.10
imageio==2.9.0
imagesize==1.2.0
immutables==0.14
importlib-metadata==1.7.0
intervaltree==3.0.2
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server==1.2.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==21.2.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.33.0+1.g022ab0f
locket==0.2.0
lxml==4.5.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
nose==1.3.7
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.19.5
numpydoc==1.1.0
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
openpyxl==3.0.4
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parse==1.19.0
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==20.1.1
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.17.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
pyodbc==4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.7.2
Rtree==0.9.4
ruamel-yaml==0.15.87
scikit-image==0.16.2
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.1
SecretStorage==3.1.2
Send2Trash==1.5.0
setuptools==49.2.0.post20200714
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.3
spyder==4.1.4
spyder-kernels==1.9.2
SQLAlchemy==1.3.18
statsmodels==0.11.1
sympy==1.6.1
tables==3.6.1
tblib==1.6.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.9
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
yapf==0.30.0
zict==2.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==4.7.1
